// src/server/fetch-clients/paystack/schemas.ts
export const baseResponseSchema = z.object({
  status: z.boolean(),
  message: z.string(),
});
// ...
export const planIntervalEnum = z.enum([
  "hourly",
  "daily",
  // "weekly",
  "monthly",
  // "quarterly",
  // "biannually",
  "annually",
]);

export const countryEnum = z.enum([
  "ghana",
  "kenya",
  "nigeria",
  "south africa",
]);

export const currencyEnum = z.enum(["XOF", "NGN", "KES", "GHS", "ZAR"]);

export type PlanInterval = z.infer<typeof planIntervalEnum>;

export const subscriptionStatusEnum = z.enum([
  "active",
  "non-renewing",
  "attention",
  "cancelled",
  "completed",
]);

export type SubscriptionStatus = z.infer<typeof subscriptionStatusEnum>;

export const planSchema = z.object({
  id: z.coerce.string(),
  name: z.string(),
  plan_code: z.string(),
  description: z.string().nullable().optional(),
  amount: z.number(),
  interval: planIntervalEnum.describe("Interval for the plan"),
  integration: z.number().optional(),
});

export const createPlanSchema = z.object({
  name: z.string().describe("Name of the plan"),
  amount: z.number().describe("Amount to be charged in subunits"),
  interval: planIntervalEnum.describe("Interval for the plan"),
  description: z.string().optional().describe("Description of the plan"),
  send_invoices: z
    .boolean()
    .optional()
    .describe(
      "Set to false if you don't want invoices to be sent to your customers",
    ),
  send_sms: z
    .boolean()
    .optional()
    .describe(
      "Set to false if you don't want text messages to be sent to your customers",
    ),
  currency: z.string().optional().describe("Currency in which amount is set"),
  invoice_limit: z
    .number()
    .optional()
    .describe("Number of invoices to raise during subscription"),
});